Replaced with long double

+++ NoteOperations.cpp
@@ -4,10 +4,10 @@
 #include <cmath>
 
 
-double midiOctaves[] = { 8.175, 16.35, 32.7, 65.4, 130.8, 261.6, 523.2,
+long double midiOctaves[] = { 8.175, 16.35, 32.7, 65.4, 130.8, 261.6, 523.2,
                          1046.4, 2092.8, 4185.6, 8371,2, 16742.4};
 
-double subContrOctave[] = { 16.35, 17.32, 18.35, 19.44, 20.61, 21.82,
+long double subContrOctave[] = { 16.35, 17.32, 18.35, 19.44, 20.61, 21.82,
                             23.12, 24.50, 25.95, 27.50, 29.13, 30.87 };
 
 std::string noteNames[] = { "C", "C#", "D", "D#", "E", "F",
@@ -31,17 +31,17 @@ std::string nameMusicInterval(const int semitones) {
 }
 
 
-double midiCents(const double f0, const double tuneFreq,
+long double midiCents(const long double f0, const long double tuneFreq,
                                       const int midiTune)
 {
-    double note = 12.0 * std::log2(f0 / tuneFreq) + midiTune;
+    long double note = 12.0 * std::log2(f0 / tuneFreq) + midiTune;
     return note * 100.0;
 }
 
 
-std::pair<int, double> findClosestNote(const double f0) {
-    const double cents = midiCents(f0);
-    const double midiNote = cents / 100.0;
+std::pair<int, long double> findClosestNote(const long double f0) {
+    const long double cents = midiCents(f0);
+    const long double midiNote = cents / 100.0;
     const long double closestNote = std::round(midiNote);
     const long double centsDeviation = (midiNote - closestNote) * 100.0;
     return {closestNote, centsDeviation};
@@ -61,10 +61,10 @@ std::string nameMidiNote(const int midiNote) {
 }
 
 
-double freqFromMidiNote(const int midiNote) {
+long double freqFromMidiNote(const int midiNote) {
     const int noteIdx = midiNote % 12;
     const int octave = (midiNote / 12) - 1;
-    double octavePow = std::pow(2.0, octave);
+    long double octavePow = std::pow(2.0, octave);
     return subContrOctave[noteIdx] * octavePow;
 }
